Make diff nicer
